+++ CHANGELOG.md
@@ -47,11 +47,14 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - `Pos`, `Motion` and `Rotation` in entity nbt are now saved as list tag instead of compound tag to match vanilla.
   This also fixed the bug that entities being spawned in incorrect position when placing structure using `/structure`
   command. Please note that this change is not backward compatible and will break the old world and player data.
-- Fixed several NaNs caused by `Vector3fc#normalize` methods in the physics engine, and now setting the motion of an
-  entity to a vector which contains NaN will result in an error.
+- Fixed several NaNs caused by `Vector3fc#normalize` methods in the physics engine, and now setting the motion/location
+  of an
+  entity to a vector which contains NaN will result in an exception.
 - EntityItem now won't have drowning damage when it is in water, this bug causes entity item died after a period of time
   in water.
 - `ServerboundLoadingScreenPacket` won't spam warnings in the console when switching dimension now.
+- Fixed the bug that sometimes there may be `NaN` values in `PlayerAuthInputPacket`, this bug is also confirmed in
+  df-mc ([issue#425](https://github.com/df-mc/dragonfly/issues/425)).
 
 ## 0.1.1 (API 0.2.0)